Add more command in Code Generation

+++ src/IdentificationTable.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public class IdentificationTable {
 
 	private int level;
-	public int lastInDMem;
+	public byte lastInDMem;
 	private LinkedList<Map<String,Attribute>> idenTable;
 	
 	public IdentificationTable(){
